+++ Backend/Api/Models/Treatment.cs
@@ -1,0 +1,5 @@
+namespace Api.Models;
+
+public class Treatment {
+
+}
